nlp-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
spacy==3.7.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
scikit-learn==1.3.2